Add release workflows for each package
These workflows assume that tags will be formed as `evo-files@v1.2.3`,
will skip if they are not prefixed with each package, and then will
publish to PyPI and upload artifacts to the GitHub release.

Individual workflow files are required for PyPI trusted publishing,
since we are using a monorepo serving multiple projects.